added TerrainMaterialManager
TerrainMaterialManager checks through all Materials on a terrain and puts them into a splatmap. it checks for the alpha of each texture on the map and uses the players coordinates to tell which texture the player is currently on. It then uses a switch to play the desired sound based on the index of the texture they're currently on.